doc_checker/serializers.py
from rest_framework import serializers

from doc_checker.models import Document


class DocumentSerializer(serializers.ModelSerializer):
    class Meta:
        model = Document
        fields = '__all__'

