requirements.txt
aiosqlite==0.17.0
annotated-types==0.6.0
asgiref==3.8.1
coverage==7.5.1
Django==5.0.6
django-smtp-ssl==1.0
djangorestframework==3.15.1
iso8601==1.1.0
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
pypika-tortoise==0.1.6
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.5.0
typing_extensions==4.11.0
